#35 refactoring login form
